requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
Flask-Cors
python-dotenv
